Extract sort logic into helper function for reusability
 - Extracted sort logic from multiple components into _getDisplaySortForColumn helper function
 - Simplified sort condition checks in sortIndicatorComp.ts and columnMenuFactory.ts
 - Added missing import for _getDisplaySortForColumn in main-internal.ts
 - Removed redundant code duplication across files
 - Improved maintainability and code consistency
